Removes an observation about non-existent case.
I believe that the case which this observation describes a handling for, is not possible.

If each aggregation window is the same size, and the timer of the later window is the same size as the timer of the earlier window… Therefore, I believe that there is no possibility of a timer from a later window reaches 0 before a timer from an earlier window.

The earlier window timer will always end before the later window timer. Even when the timer size (defined by the New Relic user) is smaller than the window aggregation size (also defined by the New Relic user).  

Does that make sense to you too, guys? (If not, I would love to understand more, please ❤️ )